[REFACTOR] Section 엔티티를 Line 애그리거트에 통합
