+++ src/test/java/nextstep/subway/line/domain/model/LineTest.java
@@ -5,10 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.springframework.boot.test.context.SpringBootTest;
 
-import nextstep.subway.section.domain.model.Distance;
-import nextstep.subway.section.domain.model.Section;
 import nextstep.subway.station.domain.model.Station;
 
 class LineTest {
